requirements.txt
cachetools==5.2.0
Flask==2.2.5
Flask_RESTful==0.3.9
PyMySQL==1.0.2
PyYAML==6.0.1
requests==2.31.0
waitress==2.1.2
xmltodict==0.13.0
Werkzeug==2.3.8
flask_swagger_ui==4.11.1
pymssql==2.2.11
python-dotenv==1.0.1
pyodbc==5.1.0